load_api_to_gcs/requirements.txt
fastapi
uvicorn
requests
google-cloud-storage
